+++ src/test/java/roomescape/time/controller/ReservationTimePageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.time.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -10,11 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.reservation.controller.ReservationPageController;
 import roomescape.reservation.service.ReservationService;
 
-@WebMvcTest(ReservationPageController.class)
-class ReservationPageControllerTest {
+@WebMvcTest(ReservationTimePageController.class)
+class ReservationTimePageControllerTest {
 
     @MockBean
     private ReservationService reservationService;
@@ -23,10 +22,10 @@ class ReservationPageControllerTest {
     private MockMvc mockMvc;
 
     @Test
-    @DisplayName("/reservation 를 요청하면 reservation.html 를 반환한다.")
-    void requestAdminReservation() throws Exception {
-        mockMvc.perform(get("/reservation"))
+    @DisplayName("/time 을 요청하면 time.html 를 반환한다.")
+    void requestTime() throws Exception {
+        mockMvc.perform(get("/time"))
                 .andExpect(status().isOk())
-                .andExpect(view().name("admin/reservation"));
+                .andExpect(view().name("admin/time"));
     }
 }
